requirements.txt
Django==2.2.9
djangorestframework==3.11.0
psycopg2>=2.7.5,<2.8.0
Markdown==3.1.1
django-filter==2.2.0